+++ lib/src/download_assets_controller.dart
@@ -31,15 +31,24 @@ abstract class DownloadAssetsController {
   /// Clear all download assets, if it already exists on local storage.
   Future clearAssets();
 
-  /// Start download of your content to local storage, uncompress all data and delete
+  /// Start the download of your content to local storage, uncompress all data and delete
   /// the compressed file.
   /// [assetsUrl] -> Specify the url for your compressed file. (http://{YOUR_DOMAIN}:{FILE_NAME}.zip
   /// [onProgress] -> It's not required. Called after each iteration returning the current progress
   /// [onComplete] -> It's not required. Called when the progress is completed with success
   /// [zippedFile] -> Zipped file's name (default value is assets.zip)
+  /// [onCancel] -> Cancel the download (optional)
+  /// [requestQueryParams] -> Query params to be used in the request (optional)
+  /// [requestExtraHeaders] -> Extra headers to be added in the request (optional)
   Future startDownload({
     required String assetsUrl,
     Function(double)? onProgress,
+    Function()? onCancel,
     String zippedFile = 'assets.zip',
+    Map<String, dynamic>? requestQueryParams,
+    Map<String, String> requestExtraHeaders = const {},
   });
+
+  /// Cancel the download
+  void cancelDownload();
 }