+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.4 - May 14, 2023
+
+* Added feature to cancel the download
+* Now allows send queryParameters and extra headers in the request
+
 ## 3.0.3 - January 27, 2023
 
 * Lint applied to project
@@ -5,9 +10,11 @@
 
 ## 3.0.0 - December 27, 2021
 
-* BREAKING REFACTOR: It's no more required the onProgress callback and was removed the onComplete callback;
+* BREAKING REFACTOR: It's no more required the onProgress callback and was removed the onComplete
+  callback;
 * All async code from the constructor was removed;
-* Removed setAssetDir method and added init method. (check de documentation or example to see how to use it)
+* Removed setAssetDir method and added init method. (check de documentation or example to see how to
+  use it)
 
 ## 2.1.0 - November 26, 2021